created add to cart function and add cart.html
created adToCart function. inside the function created an addEventListener which create an object of the cart and set item to local storage. then created a for loop to get item from local storage and append it to code.
Then created a for loop that get price items from local storage, push it on an empty array and sho the total price of the cart