developed POST request and change form btn
Developed addEventListener for form contact that create a form object to send to backend. The form object has a contact object with inside the input of each field of contact form, and product object with inside the product array. 
Then fetch the POST request to send data to backend and get orderId.